app/requirements.txt
aiogram==2.12.1
asyncpg==0.22.0
asyncpgsa==0.27.1
python-dotenv==0.17.0
aioredis==1.3.1
aiocache==0.11.1
ujson==4.0.2
aiosqlite==0.17.0
pyrogram==1.2.8
tgcrypto==1.2.2
aiofile==3.5.0
sqlalchemy==1.4.7